chanstate: move revocation tlv adapters
Move the SparsePayHash TLV adapter and HTLCEntry stream builder into the
revocation KV store file. These helpers describe the revocation log
on-disk TLV layout, so keep them with the rest of the KV serialization
while the domain file keeps the exported types and constructors.
